% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_exoweights}
\alias{check_exoweights}
\title{Checks whether the given exogenous transition weights for simulation are correctly specified.}
\usage{
check_exoweights(M, exo_weights, how_many_rows, name_of_row_number)
}
\arguments{
\item{exo_weights}{Exogenous transition weights weights given for simulation in some context.}

\item{how_many_rows}{how many rows the exogenous weights should have?}

\item{name_of_row_number}{what is the name of the object whose value should determine the
the number of rows in the exogenous weights?}
}
\value{
Throws an error if the exogenous weights are incorrectly specified.
}
\description{
\code{check_exoweights} checks whether the given exogenous transition weights for
 simulation are correctly specified.
}
\details{
Used by simulate.stvar, predict.stvar, GIRF, and GFEVD.
}
\keyword{internal}
