% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_pMd}
\alias{check_pMd}
\title{Check that p, M, and d are correctly set}
\usage{
check_pMd(
  p,
  M,
  d,
  weight_function = c("relative_dens", "logistic", "mlogit", "exponential", "threshold",
    "exogenous"),
  identification = c("reduced_form", "recursive", "heteroskedasticity",
    "non-Gaussianity")
)
}
\arguments{
\item{p}{the autoregressive order of the model}

\item{M}{the number of regimes}

\item{d}{the number of time series in the system, i.e., the dimension}
}
\value{
Throws an error if something is wrong.
}
\description{
\code{check_pMd} checks the arguments p, M, and d.
}
\keyword{internal}
