% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_stvar}
\alias{check_stvar}
\title{Checks whether the given object has class attribute 'stvar'}
\usage{
check_stvar(object, object_name)
}
\arguments{
\item{object}{S3 object to be tested}

\item{object_name}{what is the name of the object that should of class 'stvar'?}
}
\value{
Throws an error if the object doesn't have the class attribute 'stvar'.
}
\description{
\code{check_stvar} checks that the object has class attribute 'stvar'.
}
\keyword{internal}
