% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{check_weightfun_pars}
\alias{check_weightfun_pars}
\title{Check the argument \code{weightfun_pars}}
\usage{
check_weightfun_pars(
  p,
  d,
  weight_function = c("relative_dens", "logit"),
  weightfun_pars = NULL
)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order}

\item{weight_function}{what type of transition weights should be used?}

\item{weightfun_pars}{\describe{
  \item{If \code{weight_function == "relative_dens"}:}{Not used.}
  \item{If \code{weight_function == "logit"}:}{a list of two elements: \describe{
    \item{The first element \code{$vars}:}{a numeric vector containing the variables that should used as switching variables
      in the weight function, i.e., a vector with unique elements in \eqn{\lbrace 1,...,d \rbrace}.}
    \item{The second element \code{$lags}:}{a numeric vector containing the lags that should be used in the weight function, i.e.,
      a vector with unique elements in \eqn{\lbrace 1,...,p \rbrace}.}
  }}
}}
}
\value{
Does checks the argument \code{weightfun_pars} and throws an error if something is wrong; returns
  a corrected version of the argument if possible.
}
\description{
\code{check_weightfun_pars} checks that the argument \code{weightfun_pars}.
  is correctly set, if not, tries to correct them.
}
\keyword{internal}
