% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{d_lift}
\alias{d_lift}
\title{Calculate the dth induced matrix of a square matrix}
\usage{
d_lift(A)
}
\arguments{
\item{Omega}{a positive definite covariance matrix}
}
\value{
a vectorized symmetric square root matrix of the matrix \code{Omega}.
}
\description{
\code{d_lift} the dth induced matrix of a square matrix
 by so-called d-lifting.
}
\references{
reference
}
\keyword{internal}
