% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dec2binar}
\alias{dec2binar}
\title{Transform a decimal to a binary array}
\arguments{
\item{n}{the index determining the submatrix of the square matrix A the permanent is calculated for.}

\item{dim}{the number of rows (or columns) in the square matrix A the permanent is calculated for.}
}
\value{
Returns a characteric string that allows to select the submatrix.
}
\description{
Transforms a decimal to a binary array with an additional
  element saving the number of ones in the array. Helper function for
  \code{get_permanent_Cpp}.
}
\details{
Helper function for \code{get_permanent_Cpp}, which calculates the
  permanent of a square matrix using the Ryser's formula.
  This function is an Rcpp implementation of the C++ code written by "hanzzoid"
  on the "Code Project" website
  \url{https://www.codeproject.com/Articles/21282/Compute-Permanent-of-a-Matrix-with-Ryser-s-Algorit}.
  Modified code reproduced under The Code Project Open licence
  (\url{https://www.codeproject.com/info/cpol10.aspx}).
}
\keyword{internal}
