% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianDensities}
\alias{GaussianDensities}
\title{Calculate log multivariate Gaussian densities}
\usage{
GaussianDensities(const_term, obs_minus_cmean)
}
\arguments{
\item{const_term}{the constant term \code{-0.5*d*log(2*pi)} that does not vary.}

\item{obs_minus_cmean}{a \eqn{(T \times d)} matrix containing the observation minus
the (conditional) mean in each row.}
}
\value{
a numeric vector containing the multivariate Gaussian densities
}
\description{
Calculates log of multivariate Gaussian densities with varying mean and covariance matrix.
}
