% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianDensities2}
\alias{GaussianDensities2}
\title{Calculate log multivariate Gaussian densities}
\usage{
GaussianDensities2(const_term, obs_minus_mean, covmat)
}
\arguments{
\item{const_term}{the constant term \code{-0.5*d*log(2*pi)} that does not vary.}

\item{obs_minus_mean}{a \eqn{(T \times dp)} such that the i:th row of the matrix contains
the vector \eqn{(y_{i-1}',...,y_{i-p}') - }\code{rep(all_mu[,m], times=p)}
\eqn{((dp)x1)}, where \eqn{y_{i}=(y_{1i},...,y_{di})} \eqn{(dx1)}. That is, the initial values are
included but the last observations not (used in relative dens transition weights).}

\item{covmat}{the \eqn{(dp \times dp)} covariance matrix that is the same for all
observations.}
}
\value{
a numeric vector containing the multivariate Gaussian densities
}
\description{
Calculates logs of multivariate Gaussian densities with varying mean but
  constant covariance matrix.
}
