% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterFactuals.R
\name{get_B_yt}
\alias{get_B_yt}
\title{Compute the impact matrix \eqn{B_{y,t}=\sum_{m=1}^M\alpha_{m,t}B_m} for a single time period}
\usage{
get_B_yt(all_Bm, alpha_mt)
}
\arguments{
\item{all_Bm}{a 3D array such that the impact matrix of \eqn{m}th regime is obtained from \code{all_Bm[, , m]}.}

\item{alpha_mt}{an \eqn{(M \times 1)} vector containing the time period \eqn{t} transition weights.}
}
\value{
Returns the \eqn{(d \times d)} impact matrix for the time period \eqn{t}.
}
\description{
\code{get_B_yt} computes the conditional mean \eqn{B_{y,t}=\sum_{m=1}^M\alpha_{m,t}B_m} for a single time period
 based on the impact matrices of the regimes and the vector of transition weights.
}
\details{
This is used in simulation of the counterfactual scenarios.
}
\keyword{internal}
