% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigens.R
\name{get_boldA_eigens}
\alias{get_boldA_eigens}
\title{Calculate absolute values of the eigenvalues of the "bold A" matrices containing the AR coefficients}
\usage{
get_boldA_eigens(stvar)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}
}
\value{
Returns a matrix with \eqn{d*p} rows and \eqn{M} columns - one column for each regime.
 The \eqn{m}th column contains the absolute values (or modulus) of the eigenvalues of the "bold A" matrix containing
 the AR coefficients correspinding to regime \eqn{m}.
}
\description{
\code{get_boldA_eigens} calculates absolute values of the eigenvalues of
  the "bold A" matrices containing the AR coefficients for each regime.
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis, \emph{Springer}.
 }
}
\keyword{internal}
