% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{get_IC}
\alias{get_IC}
\title{Calculate AIC, HQIC, and BIC}
\usage{
get_IC(loglik, npars, T_obs)
}
\arguments{
\item{loglik}{log-likelihood value, preferably non-penalized one.}

\item{npars}{number of (freely estimated) parameters in the model}

\item{T_obs}{numbers of observations with the \eqn{p} starting values excluded.}
}
\value{
Returns a data frame containing the information criteria values
  divided by the number of observations.
}
\description{
\code{get_IC} calculates the information criteria values
  AIC, HQIC, and BIC divided by the number of observations.
}
\keyword{internal}
