% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{get_new_start}
\alias{get_new_start}
\title{Get the new starting time of series that is forwarded some number of steps}
\usage{
get_new_start(y_start, y_freq, steps_forward)
}
\arguments{
\item{y_start}{original starting time of the series}

\item{y_freq}{frequency of the series}

\item{steps_forward}{how many steps the series should be forwarded?}
}
\value{
Returns a length two numeric vector with the "year" (or "major")
  time point in the first element the "quarter/month/week/day" (or "minor")
  time in the second element for a series that is forwarded from \code{y_start}
  \code{steps_forward} steps forward.
}
\description{
\code{get_new_start} calculates the new starting time of series
  that is forwarded some number of steps.
}
\keyword{internal}
