% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigens.R
\name{get_omega_eigens}
\alias{get_omega_eigens}
\title{Calculate the eigenvalues of the "Omega" error term covariance matrices}
\usage{
get_omega_eigens(stvar)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}
}
\value{
Returns a matrix with \eqn{d} rows and \eqn{M} columns - one column for each regime.
 The \eqn{m}th column contains the eigenvalues of the "Omega" error term covariance matrix
 of the \eqn{m}th regime.
}
\description{
\code{get_omega_eigens} calculates the eigenvalues of the "Omega" error
 term covariance matrices for each regime
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis, \emph{Springer}.
 }
}
\keyword{internal}
