% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_permanent_Cpp}
\alias{get_permanent_Cpp}
\title{Calculate the permanent of a matrix}
\usage{
get_permanent_Cpp(A)
}
\arguments{
\item{A}{a square matrix}
}
\value{
Returns the permanent of the matrix A (a real number)
}
\description{
Calculates the permanent of a square matrix using the Ryser's formula.
}
\details{
Calculates the permanent of a square matrix using Ryser's formula.
}
\keyword{internal}
