% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIRFandGFEVD.R
\name{GFEVD}
\alias{GFEVD}
\title{Estimate generalized forecast error variance decomposition for structural
 STVAR models.}
\usage{
GFEVD(
  stvar,
  shock_size = 1,
  N = 30,
  initval_type = c("data", "random", "fixed"),
  R1 = 250,
  R2 = 250,
  init_regime = 1,
  init_values = NULL,
  which_cumulative = numeric(0),
  ncores = 2,
  burn_in = 1000,
  seeds = NULL
)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'}, created by, e.g., \code{fitSTVAR} or \code{fitSSTVAR}.}

\item{shock_size}{What shocks size should be used for all shocks? By the normalization, the conditional
covariance matrix of the structural error is an identity matrix.}

\item{N}{a positive integer specifying the horizon how far ahead should the GFEVD be calculated.}

\item{initval_type}{What type initial values are used for estimating the GIRFs that the GFEVD is based on?
\describe{
  \item{\code{"data"}:}{Estimate the GIRF for all the possible length \eqn{p} histories in the data.}
  \item{\code{"random"}:}{Estimate the GIRF for several random initial values generated from the a specific regime
     specified by the argument \code{init_regimes}. The number of initial values is set with the argument \code{R2}.}
  \item{\code{"fixed"}:}{Estimate the GIRF for a fixed initial value only, which is specified with the argument
     \code{init_values}.}
}}

\item{R1}{the number of repetitions used to estimate GIRF for each initial
value.}

\item{R2}{the number of initial values to be drawn if \code{initval_type="random"}.}

\item{init_regime}{an integer in \eqn{1,...,M} specifying the regime from which
the initial values should be generated from. The initial values will be generated
from the stationary distribution of the specific regime. Due to the lack of
knowledge of the stationary distribution, models with other than Gaussian conditional distribution
use a simulation procedure with a burn-in period. See the details section.}

\item{init_values}{a size \eqn{(pxd)} matrix specifying the initial values, where d is the number
of time series in the system. The \strong{last} row will be used as initial values for the first lag,
the second last row for second lag etc. If not specified, initial values will be drawn from
the regime specified in \code{init_regimes} (for Gaussian models only).}

\item{which_cumulative}{a numeric vector with values in \eqn{1,...,d}
(\code{d=ncol(data)}) specifying which the variables for which the impulse
responses should be cumulative. Default is none.}

\item{ncores}{the number CPU cores to be used in parallel computing. Only
single core computing is supported if an initial value is specified (and
the GIRF won't thus be estimated multiple times).}

\item{seeds}{a numeric vector containing the random number generator seed for estimation
of each GIRF. Should have the length...
\itemize{
  \item ...\code{nrow(data) - p + 1} if \code{initval_type="data"}.
  \item ...\code{R2} if \code{initval_type="random"}.
  \item ...\code{1} if \code{initval_type="fixed."}.
}
Set to \code{NULL} for not initializing the seed. Exists for creating reproducible results.}
}
\value{
Returns and object of class 'gfevd' containing the GFEVD for all the variables and to
  the transition weights. Note that the decomposition does not exist at horizon zero for transition weights
  because the related GIRFs are always zero at impact.
}
\description{
\code{GFEVD} estimates generalized forecast error variance decomposition
 for structural STVAR models.
}
\details{
The GFEVD is a forecast error variance decomposition calculated with the generalized
  impulse response function (GIRF). See Lanne and Nyberg (2016) for details.
}
\examples{
 \donttest{
 # These are long-running examples that use parallel computing.
 # It takes approximately 30 seconds to run all the below examples.

 ## FILL IN
 }
}
\references{
\itemize{
   \item Lanne M. and Nyberg H. 2016. Generalized Forecast Error Variance Decomposition for Linear
     and Nonlineae Multivariate Models. \emph{Oxford Bulletin of Economics and Statistics}, \strong{78}, 4, 595-603.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{linear_IRF}}, \code{\link{fitSSTVAR}}
}
