% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{order_B}
\alias{order_B}
\title{Reorder columns of a square matrix so that the first nonzero elements are
 in decreasing order}
\usage{
order_B(B)
}
\arguments{
\item{B}{A square numeric matrix.}
}
\value{
A square matrix \code{B} with columns reordered so that its
diagonal entries are in a decreasing order.
}
\description{
\code{order_B} takes a square matrix \code{B} as input and reorders its columns
so that the diagonal entries of \code{B} are in decreasing order. The function
is designed to be computationally efficient and ensures that the input is a square matrix.
}
\keyword{internal}
