% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagTests.R
\name{Portmanteau_test}
\alias{Portmanteau_test}
\title{Perform Portmanteau test for a STVAR model}
\usage{
Portmanteau_test(stvar)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'} generated by \code{fitSTVAR} or \code{STVAR}.}
}
\value{
A list with class "hypotest" containing the test results and arguments used to calculate the test.
}
\description{
\code{Portmanteau_test} performs Portmanteau test for remaining autocorrealtion
  in the residuals  for a STVAR model
}
\details{
The implemented test is discussed, for instance, in Lütkepohl (2005), Section 4.4.3.
}
\examples{
# FILL IN
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
           \emph{Springer}.
  }
}
\seealso{
\code{\link{LR_test}}, \code{\link{Rao_test}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
  \code{\link{diagnostic_plot}}, \code{\link{profile_logliks}},
}
