% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.histdecomp}
\alias{print.histdecomp}
\title{Print method for the class histdecomp}
\usage{
\method{print}{histdecomp}(x, ..., digits = 3, which_vars, which_indices)
}
\arguments{
\item{x}{object of class \code{'histdecomp'} generated by the function \code{hist_decomp}}

\item{...}{currently not in use.}

\item{digits}{how many significant digits to print?}

\item{which_vars}{a numeric vector specifying the variables to print. The default is that
all the variables are printed.}

\item{which_indices}{a numeric vector specifying the time period indices to print. The default
is that all the time periods are printed.}
}
\value{
Returns the input object \code{x} invisibly.
}
\description{
\code{print.hypotest} is the print method for the class histdecomp
 objects.
}
