% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.stvarsum}
\alias{print.stvarsum}
\title{Summary print method from objects of class 'stvarsum'}
\usage{
\method{print}{stvarsum}(x, ..., digits, standard_error_print = FALSE)
}
\arguments{
\item{x}{object of class 'stvarsum' generated by \code{summary.stvar}.}

\item{...}{currently not used.}

\item{digits}{the number of digits to be printed.}

\item{standard_error_print}{if set to \code{TRUE}, instead of printing the estimates,
prints the approximate standard errors using square roots of the diagonal of inverse
of the observed information matrix.}
}
\value{
Returns the input object \code{x} invisibly.
}
\description{
\code{print.stvarsum} is a print method for object \code{'stvarsum'} generated
  by \code{summary.stvar}.
}
