% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrors.R
\name{print_std_errors}
\alias{print_std_errors}
\title{Print standard errors of a smooth transition VAR model in the same form as the model estimates are printed}
\usage{
print_std_errors(stvar, digits = 3)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{digits}{number of digits to be printed.}
}
\description{
\code{print_std_errors} prints the approximate standard errors of a smooth transition VAR model in the
  same form as the parameters of objects of class \code{'stvar'} are printed.
}
\details{
\strong{Note that the approximate standard errors are based on the unverified assumption of
  asymptotic normality of the estimator!}

  The main purpose of \code{print_std_errors} is to provide a convenient tool to match the standard
  errors to certain parameter estimates. Note that if the model is intercept parametrized, there won't
  be standard errors for the unconditional means, and vice versa.

  Note that if linear constraints are imposed and they involve summations or multiplications, then the AR
  parameter standard errors are printed separately as they don't correspond one-to-one to the model parameter
  standard errors.
}
\examples{
\donttest{
# STVAR p=1, M=2 models
mod12 <- fitSTVAR(gdpdef, p=1, M=2, nrounds=1, seeds=4)
mod12
print_std_errors(mod12)
}
}
\references{
\itemize{
   \item TO BE FILLED IN
 }
}
\seealso{
\code{\link{fitSTVAR}}, \code{\link{STVAR}}, \code{\link{print.stvar}},
 \code{\link{swap_parametrization}}
}
