% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{random_df}
\alias{random_df}
\title{Create random degrees of freedom parameter values}
\usage{
random_df(how_many)
}
\arguments{
\item{how_many}{how many parameter values should be drawn?}
}
\value{
Returns a numeric vector of length \code{how_many} with random entries strictly larger than two.
}
\description{
\code{random_df} generates random degrees of freedom parameter values
}
\keyword{internal}
