% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{random_distpars}
\alias{random_distpars}
\title{Create random distribution parameter values}
\usage{
random_distpars(cond_dist)
}
\arguments{
\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"} or \code{"Student"}.}
}
\value{
Returns a numeric vector ...
  \describe{
    \item{If \code{cond_dist == "Gaussian"}:}{of length zero.}
    \item{If \code{cond_dist == "Student"}:}{of length one containing a df param strictly larger than two.}
  }
}
\description{
\code{random_distpars} generates random distribution parameter values
}
\keyword{internal}
