% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STVARconstruction.R
\name{reorder_W_columns}
\alias{reorder_W_columns}
\title{Reorder columns of the W-matrix and lambda parameters of a structural STVAR model
  that is identified by heteroskedasticity.}
\usage{
reorder_W_columns(stvar, perm, calc_std_errors = FALSE)
}
\arguments{
\item{stvar}{a class 'stvar' object defining a structural STVAR model that is identified by heteroskedasticity,
typically created with \code{fitSSTVAR}.}

\item{perm}{an integer vector of length \eqn{d} specifying the new order of the columns of \eqn{W}.
Also lambda parameters of each regime will be reordered accordingly.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining a structural STVAR model with the columns of \eqn{W}
  reordered.
}
\description{
\code{reorder_W_columns} reorder columns of the W-matrix and lambda parameters
  of a structural STVAR model that is identified by heteroskedasticity.
}
\details{
The order of the columns of \eqn{W} can be changed without changing the implied reduced
  form model as long as the order of lambda parameters is also changed accordingly. Note that the
  constraints imposed on \eqn{W} (or the  impact matrix) will also be modified accordingly.

  This function does not support models with constraints imposed on the lambda parameters!

  Also all signs in any column of \eqn{W} can be swapped (without changing the implied reduced form model)
  with the function \code{swap_W_signs} but this obviously also swaps the sign constraints in the
  corresponding columns of \eqn{W}.
}
\examples{
# p=2, M=2, d=2, Student's t logistic STVAR model with the first lag of the second
 # variable  as the switching variable and shocks identified by heteroskedasticity.
theta_222logt <- c(0.356914, 0.107436, 0.356386, 0.086330, 0.139960, 0.035172, -0.164575,
  0.386816, 0.451675, 0.013086, 0.227882, 0.336084, 0.239257, 0.024173, -0.021209, 0.707502,
  0.063322, 0.027287, 0.009182, 0.197066, -0.03, 0.24, -0.76, -0.02, 3.36, 0.86, 0.1, 0.2, 7)
mod222logt <- STVAR(data=gdpdef, p=2, M=2, d=2, params=theta_222logt,
                               weight_function="logistic", weightfun_pars=c(2, 1),
                               cond_dist="Student", identification="heteroskedasticity")
mod222logt

# Reverse the ordering of the columns of W:
mod222logt_rev <- reorder_W_columns(mod222logt, perm=c(2, 1))
mod222logt_rev

# Swap the ordering back to the original:
mod222logt_rev2 <- reorder_W_columns(mod222logt_rev, perm=c(2, 1))
mod222logt_rev2

# Does not do anything, as perm=1:2 so the ordering does not change:
reorder_W_columns(mod222logt, perm=c(1, 2))
}
\references{
\itemize{
   \item Lütkepohl H., Netšunajev A. 2018. Structural vector autoregressions with smooth transition in variances.
     \emph{Journal of Economic Dynamics & Control}, \strong{84}, 43-57.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{fitSSTVAR}}, \code{\link{swap_W_signs}}
}
