% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{smart_df}
\alias{smart_df}
\title{Create random degrees of freedom parameter values close to given values}
\usage{
smart_df(df, accuracy)
}
\arguments{
\item{df}{the old degrees of freedom parameters (of all regimes)}

\item{accuracy}{a positive real number adjusting how close to the given degrees of freedom parameters
the returned df should be.}
}
\value{
Returns a numeric vector of length \code{how_many} with random entries strictly larger than two.
}
\description{
\code{random_df} generates random \code{M2} degrees of freedom parameter values
  close to given values, where \code{M2} is number of StMVAR type regimes in the model.
}
\keyword{internal}
