% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{smart_weightpars}
\alias{smart_weightpars}
\title{Create random transition weight parameter values}
\usage{
smart_weightpars(M, weight_pars, weight_function, accuracy)
}
\arguments{
\item{M}{a positive integer specifying the number of regimes}

\item{weight_pars}{a vector containing transition weight parameter values.
\describe{
  \item{If \code{weight_function == "relative_dens"}:}{a length \code{M-1} vector \eqn{(\alpha_1,...,\alpha_{M-1})}.}
  \item{If \code{weight_function == "logit"}:}{NOT YET IMPLEMENTED}
}}

\item{weight_function}{what type of transition weights should be used? See the vignette for details about the weight functions.}

\item{accuracy}{a positive real number adjusting how close to the given parameter vector the returned individual should be.
Larger number means larger accuracy. Read the source code for details.}
}
\value{
Returns a numeric vector ...
  \describe{
    \item{If \code{weight_function == "relative_dens"}:}{a length \code{M-1} vector \eqn{(\alpha_1,...,\alpha_{M-1})}.}
    \item{If \code{weight_function == "logit"}:}{NOT YET IMPLEMENTED}
  }
}
\description{
\code{smart_weightpars} generates random transition weight parameter values
 relatively close to the ones given in \code{weight_pars}
}
\keyword{internal}
