% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STVARconstruction.R
\name{swap_W_signs}
\alias{swap_W_signs}
\title{Swap all signs in pointed columns a the \eqn{W} matrix of a structural STVAR model
  that is identified by heteroskedasticity.}
\usage{
swap_W_signs(stvar, which_to_swap, calc_std_errors = FALSE)
}
\arguments{
\item{stvar}{a class 'stvar' object defining a structural STVAR model that is identified by heteroskedasticity,
typically created with \code{fitSSTVAR}.}

\item{which_to_swap}{a numeric vector of length at most \eqn{d} and elemnts in \eqn{1,..,d}
specifying the columns of \eqn{W} whose sign should be swapped.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining a structural STVAR model with the signs of the
  columns of \eqn{W} swapped.
}
\description{
\code{swap_W_signs} swaps all signs in pointed columns a the \eqn{W} matrix
 a structural STVAR model that is identified by heteroskedasticity. Consequently, signs in
  the columns of the impact matrix are also swapped accordingly.
}
\details{
All signs in any column of \eqn{W} can be swapped without changing the implied reduced form model.
  Consequently, also the signs in the columns of the impact matrix are swapped. Note that the sign constraints
  imposed on \eqn{W} (or the impact matrix) are also swapped in the corresponding columns accordingly.

  Also the order of the columns of \eqn{W} can be changed (without changing the implied reduced
  form model) as long as the order of lambda parameters is also changed accordingly. This can be
  done with the function \code{reorder_W_columns}.
}
\examples{
# p=2, M=2, d=2, Student's t logistic STVAR model with the first lag of the second
# variable  as the switching variable and shocks identified by heteroskedasticity.
theta_222logt <- c(0.356914, 0.107436, 0.356386, 0.086330, 0.139960, 0.035172, -0.164575,
  0.386816, 0.451675, 0.013086, 0.227882, 0.336084, 0.239257, 0.024173, -0.021209, 0.707502,
  0.063322, 0.027287, 0.009182, 0.197066, -0.03, 0.24, -0.76, -0.02, 3.36, 0.86, 0.1, 0.2, 7)
mod222logt <- STVAR(data=gdpdef, p=2, M=2, d=2, params=theta_222logt,
                               weight_function="logistic", weightfun_pars=c(2, 1),
                               cond_dist="Student", identification="heteroskedasticity")
mod222logt

# Swap the signs of the first column of W:
mod222logt2 <- swap_W_signs(mod222logt, which_to_swap=1)
mod222logt2

# Swap the signs of the second column of W:
mod222logt3 <- swap_W_signs(mod222logt, which_to_swap=2)
mod222logt3

# Swap the signs of both columns of W:
mod222logt4 <- swap_W_signs(mod222logt, which_to_swap=1:2)
mod222logt4
}
\references{
\itemize{
   \item Lütkepohl H., Netšunajev A. 2018. Structural vector autoregressions with smooth transition in variances.
     \emp{Journal of Economic Dynamics & Control}, \strong{84}, 43-57.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{fitSSTVAR}}, \code{\link{reorder_W_columns}}
}
