% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{unBvec}
\alias{unBvec}
\title{Reverse vectorization operator that restores zeros}
\usage{
unBvec(Bvector, d, B_constraints)
}
\arguments{
\item{Bvector}{a length \eqn{d^2 - n_zeros} vector where \eqn{n_zeros} is the
number of zero entries in the impact matrix.}

\item{d}{the number of rows in the square matrix to be formed.}
}
\value{
a \eqn{(d x d)} impact matrix \eqn{B}.
}
\description{
\code{unBvec} forms a square matrix from a vector of
  stacked columns where zeros are removed according to impact
  matrix constraints.
}
\section{Warning}{

 No argument checks!
}

\keyword{internal}
