% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcal.R
\name{unWvec}
\alias{unWvec}
\title{Reverse vectorization operator that restores zeros}
\usage{
unWvec(Wvector, d, B_constraints)
}
\arguments{
\item{Wvector}{a length \eqn{d^2 - n_zeros} vector where \eqn{n_zeros} is the
number of zero entries in the impact matrix.}

\item{d}{the number of rows in the square matrix to be formed.}
}
\value{
a \eqn{(d x d)} impact matrix \eqn{W}.
}
\description{
\code{unWvec} forms a square matrix from a vector of
  stacked columns where zeros are removed according to impact
  matrix constraints.
}
\keyword{internal}
