% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STVARconstruction.R
\name{alt_stvar}
\alias{alt_stvar}
\title{Construct a STVAR model based on results from an arbitrary estimation round of \code{fitSTVAR}}
\usage{
alt_stvar(stvar, which_largest = 1, which_round, calc_std_errors = TRUE)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{which_largest}{based on estimation round with which largest log-likelihood should the model be constructed?
An integer value in 1,...,\code{nrounds}. For example, \code{which_largest=2} would take the second largest log-likelihood
and construct the model based on the corresponding estimates.}

\item{which_round}{based on which estimation round should the model be constructed? An integer value in 1,...,\code{nrounds}.
If specified, then \code{which_largest} is ignored.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining the specified reduced form or structural
 smooth transition VAR model. Can be used to work with other functions provided in \code{sstvars}.
}
\description{
\code{alt_stvar} constructs a STVAR model based on results from an arbitrary estimation
  round of \code{fitSTVAR}
}
\details{
It's sometimes useful to examine other estimates than the one with the highest log-likelihood. This function
  is wrapper around \code{STVAR} that picks the correct estimates from an object returned by \code{fitSTVAR}.
}
\examples{
\donttest{
## These are long-running examples that take approximately 10 seconds to run.

# Estimate a Gaussian STVAR p=1, M=2 model with exponential weight function and
# the first lag of the second variable as the switching variables. Run only two
# estimation rounds:
fit12 <- fitSTVAR(gdpdef, p=1, M=2, weight_function="exponential", weightfun_pars=c(2, 1),
 nrounds=2, seeds=c(1, 7))
fit12$loglik # Log-likelihood of the estimated model

# Print the log-likelihood obtained from each estimation round:
fit12$all_logliks

# Construct the model based on the second largest log-likelihood found in the
# estimation procedure:
fit12_alt <- alt_stvar(fit12, which_largest=2, calc_std_errors=FALSE)
fit12_alt$loglik # Log-likelihood of the alternative solution

# Construct a model based on a specific estimation round, the second round:
fit12_alt2 <- alt_stvar(fit12, which_round=2, calc_std_errors=FALSE)
fit12_alt2$loglik # Log-likelihood of the alternative solution
}
}
\references{
\itemize{
   \item Anderson H., Vahid F. 1998. Testing multiple equation systems for common nonlinear components.
     \emph{Journal of Econometrics}, \strong{84}:1, 1-36.
   \item Hubrich K., Teräsvirta. T. 2013. Thresholds and Smooth Transitions in Vector Autoregressive Models.
     \emph{CREATES Research Paper 2013-18, Aarhus University.}
   \item Kheifets I.L., Saikkonen P.J. 2020. Stationarity and ergodicity of Vector STAR models.
     \emph{Econometric Reviews}, \strong{39}:4, 407-414.
   \item Lütkepohl H., Netšunajev A. 2017. Structural vector autoregressions with smooth transition in variances.
     \emph{Journal of Economic Dynamics & Control}, \strong{84}, 43-57.
   \item Tsay R. 1998. Testing and Modeling Multivariate Threshold Models.
     \emph{Journal of the American Statistical Association}, \strong{93}:443, 1188-1202.
 }
}
\seealso{
\code{\link{STVAR}}
}
