% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSpectralRadius.R
\name{bound_jsr}
\alias{bound_jsr}
\title{Calculate upper bound for the joint spectral radius of the "bold A" matrices}
\usage{
bound_jsr(all_boldA, accuracy = c("0.707", "0.840", "0.917", "0.957", "0.978"))
}
\arguments{
\item{all_boldA}{all \eqn{((dp)x(dp))} "bold A" (companion form) matrices in a 3D array,
so that \code{[, , m]} gives the matrix the regime \code{m}.}

\item{accuracy}{what should the relative accuracy of the bounds be? Note only upper bound is used
and that the bound holds with any accuracy (but the bounds get tighter with increased accuracy).}
}
\value{
Returns the dth induced matrix of A: a square matrix with \code{choose(nrow(A) + d - 1, d)}
  rows and columns.
}
\description{
\code{bound_JSR} calculates an upper bound for the joint spectral radius of the "bold A" matrices
as described in Parrilo and Jadbabaie (2008), Theorems 4.2 and 4.3 and Equations (8) and (11).
}
\details{
Upper bound calculated the formula presented in Parrilo and Jadbabaie (2008), Equation~(11),
 whereas the accuracies are based on  Parrilo and Jadbabaie (2008), Table~1.
}
\references{
\itemize{
   \item P.A. Parrilo, A. Jadbabaie. 2008. Approximation of the joint spectral
      radius using sum of squares. \emph{Linear Algebra and its Applications},
      \strong{428}, 2385-2402.
   \item R. Jungers (2023). The JSR toolbox (https://www.mathworks.com/matlabcentral/fileexchange/33202-the-jsr-toolbox),
      MATLAB Central File Exchange.
 }
}
\keyword{internal}
