% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSpectralRadius.R
\name{bound_JSR}
\alias{bound_JSR}
\title{Calculate upper bound for the joint spectral radius of the "companion form AR matrices" of the regimes}
\usage{
bound_JSR(
  stvar,
  epsilon = 0.01,
  print_progress = TRUE,
  method = c("Gripenberg", "JP"),
  JP_accuracy = c("0.707", "0.840", "0.917", "0.957", "0.978")
)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{epsilon}{a strictly positive real number that approximately defines the length of the interval between the lower
and upper bounds in Gripenberg's method. A smaller epsilon value results in a narrower interval, thus providing better
accuracy for the bounds, but at the cost of increased computational effort.}

\item{print_progress}{logical: should the progress of the Gripenberg's algorithm be printed?}

\item{method}{which method should be used for bounding the joint spectral radius? Gripenberg's method is highly recommended
over JP (see the details section).}

\item{JP_accuracy}{what should the relative accuracy of the upper bound of Jadabaie and Parrilo be (for \code{method="JP"})?}
}
\value{
Returns lower and upper bounds for the joint spectral radius of the "companion form AR matrices" of the regimes.
  If the upper bound by Jadbabaie and Parrilo (2008) is calculated, only the upper bound is returned.
}
\description{
\code{bound_JSR} calculates an bounds for the joint spectral radius of the
 "companion form AR matrices" matrices of the regimes to assess the validity of the stationarity condition.
}
\details{
A sufficient condition for ergodic stationarity of the STVAR processes implemented in \code{sstvars} is that the joint
 spectral radius of the "companion form AR matrices" of the regimes is smaller than one (Kheifets and Saikkonen, 2020). This function
 calculates an upper (and lower) bound for the JSR and is implemented to assess the validity of this condition in practice. If the
 bound is smaller than one, the model is deemed ergodic stationary.

 Currently, two methods are implemented: the branch-and-bound method by Gripenberg (1996) and the upper bound by
 Jadbabaie and Parrilo (2008). We highly recommend using the Gripenberg's method, as it is mainly much faster and less prone to
 memory issues than the upper bound by Jadbabaie and Parrilo (2008). Calculation of the latter with good enough accuracy may not be
 feasible for other than very small models. However, for large models also Gripenberg's method may take very long if tight bounds
 are required. When \code{print_progress == TRUE}, the tightest bounds found so-far are printed in each iteration of Gripenberg's
 algorithm, so you can also just terminate the algorithm when the bounds are tight enough for your purposes. Consider also adjusting
 the argument \code{epsilon}, as larger epsilon does not just make the bounds less tight but also speeds up the algorithm significantly.

 Various methods for bounding the JSR are discussed and compared in Chang and Blondel (2013).
}
\examples{
# p=1, M=2, d=2, relative dens weight function
theta_122relg <- c(0.734054, 0.225598, 0.705744, 0.187897, 0.259626, -0.000863,
  -0.3124, 0.505251, 0.298483, 0.030096, -0.176925, 0.838898, 0.310863, 0.007512,
  0.018244, 0.949533, -0.016941, 0.121403, 0.573269)
mod122 <- STVAR(data=gdpdef, p=1, M=2, params=theta_122relg)
# Absolute values of the eigenvalues of the "companion form AR matrices":
summary(mod122)$abs_boldA_eigens
# It is a necessary (but not sufficient!) condition for ergodic stationary that
# the spectral radius of the "companion form AR matrices" are smaller than one
# for all of the regimes. A sufficient (but not necessary) condition for
# ergodic stationary is that the joint spectral radius of the companion form
# AR matrices" of the regimes is smaller than one. Therefore, we calculate
# an upper (and lower) bound for the joint spectral radius.

## Bounds by Gripenberg's method (default and recommonded).
# Since the largest modulus of the companion form AR matrices is not very close
# to one, we likely won't need very thight bounds to verify the JSR is smaller
# than one. Thus, we set epsilon=0.01 so that the interval between the lower
# and upper bound is roughly 0.01:
bound_JSR(mod122, epsilon=0.01, method="Gripenberg")
# The upper bound is smaller than one, so the model is ergodic stationary.

# If we want tighter bounds, we can set smaller epsilon, e.g., epsilon=0.001:
bound_JSR(mod122, epsilon=0.001, method="Gripenberg")

## Upper bound by the method of Jadbabaie and Parrilo (2008):
# We start with the lowest relative accuracy to see if it is already enough
# to verify that the JSR is smaller than one:
bound_JSR(mod122, method="JP", JP_accuracy="0.707")
# accuracy "0.707" gives a rough approximation, but the upper bound is already
# smaller than one, so the sufficient condition for ergodic stationarity is
# satisfied.

# Higher accuracy gives tighter upper bound:
bound_JSR(mod122, method="JP", JP_accuracy="0.840")
bound_JSR(mod122, method="JP", JP_accuracy="0.917")
# bound_JSR(mod122, method="JP", JP_accuracy="0.957") # Takes a while to compute!
}
\references{
\itemize{
 \item C-T Chang and V.D. Blondel. 2013 . An experimental study of approximation algorithms for the joint spectral radius.
     \emph{Numerical algorithms}, \strong{64}, 181-202.
   \item Gripenberg, G. 1996. Computing the joint spectral radius. \emph{Linear Algebra and its Applications},
     234, 43–60.
   \item I.L. Kheifets, P.J. Saikkonen. 2020. Stationarity and ergodicity of Vector STAR models.
     \emph{Econometric Reviews}, \strong{39}:4, 407-414.
   \item P.A. Parrilo, A. Jadbabaie. 2008. Approximation of the joint spectral
      radius using sum of squares. \emph{Linear Algebra and its Applications}, \strong{428}, 2385-2402.
 }
}
