% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSpectralRadius.R
\name{bound_JSR}
\alias{bound_JSR}
\title{Calculate upper bound for the joint spectral radius of the "companion form AR matrices" of the regimes}
\usage{
bound_JSR(stvar, accuracy = c("0.707", "0.840", "0.917", "0.957", "0.978"))
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{accuracy}{what should the relative accuracy of the bounds be? Note only upper bound is used
and that the bound holds with any accuracy (but the bounds get tighter with increased accuracy).}
}
\description{
\code{bound_JSR} calculates an upper bound for the joint spectral radius of the
 "companion form AR matrices" matrices of the regimes as described in Parrilo and Jadbabaie (2008),
  Theorems 4.2 and 4.3 and Equations (8) and (11).
}
\examples{
# p=1, M=2, d=2, relative dens weight function
theta_122relg <- c(0.734054, 0.225598, 0.705744, 0.187897, 0.259626, -0.000863,
  -0.3124, 0.505251, 0.298483, 0.030096, -0.176925, 0.838898, 0.310863, 0.007512,
  0.018244, 0.949533, -0.016941, 0.121403, 0.573269)
mod122 <- STVAR(data=gdpdef, p=1, M=2, params=theta_122relg)
# Absolute values of the eigenvalues of the "companion form AR matrices":
summary(mod122)$abs_boldA_eigens
# It is a necessary (but not sufficient!) condition for ergodic stationary that
# the spectral radius of the "companion form AR matrices" are smaller than one
# for all of the regimes. A sufficient (but not necessary) condition for
# ergodic stationary is that the joint spectral radius of the #companion form
# AR matrices" of the regimes is smaller than one. Therefore, we calculate
# upper bounds for the joint spectral radius.

# Upper bound for the joint spectral radius of the "companion form AR matrices":
bound_JSR(mod122, accuracy="0.707")
# accuracy "0.707" gives a rough approximation, but the upper bound is already
# smaller than one, so the sufficient condition for ergodic stationarity is satisfied.

# Higher accuracy gives tighter upper bound:
bound_JSR(mod122, accuracy="0.840")
bound_JSR(mod122, accuracy="0.917")
# bound_JSR(mod122, accuracy="0.957") # Takes a while to compute!
# bound_JSR(mod122, accuracy="0.978") # Takes much longer to compute!
}
