% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSpectralRadius.R
\name{bound_jsr_G}
\alias{bound_jsr_G}
\title{Calculate upper bound for the joint spectral radius of a set of matrices}
\usage{
bound_jsr_G(S, epsilon = 0.01, print_progress = TRUE)
}
\arguments{
\item{S}{the set of matrices the bounds should be calculated for in an array, in VAR applications,
all \eqn{((dp)x(dp))} "bold A" (companion form) matrices in a 3D array, so that \code{[, , m]} gives the matrix
the regime \code{m}.}

\item{epsilon}{a strictly positive real number specifying the absolute accuracy of the bounds (which will be approximate
in practice). Smaller number gives better accuracy but requires more substantial computational effort.}
}
\value{
Returns an upper bound for the joint spectral radius of the "companion form AR matrices" of the regimes.
}
\description{
\code{bound_jsr_G} calculates lowr and upper bounds for the joint spectral radious of a set of square matrices,
 typically the "bold A" matrices, using the algorithm by Gripenberg (1996)
}
\details{
The bounds are calculated using the Gripenberg's (1996) branch-and-bound method, which is also discussed
 in Chand and Blondel (2013). Specifically, Kheifets and Saikkonen (2020) show that if the joint spectral radius
 of the companion form AR matrices of the regimes is smaller than one, the STVAR process is ergodic stationary. Therefore,
 if the upper bound is smaller than one, the process is stationary ergodic. However, as the condition is not
 necessary but sufficient and also because the bound might be too conservative, upper bound larger than one
 does not imply that the process is not ergodic stationary. You can try higher accuracy, and if the bound is
 still larger than one, the result does not tell whether the process is ergodic stationary or not.

 Note that with high precision (small \code{epsilon}), the computational effort required are substantial and
 the estimation may take very long.

 You can also try other implementations for bounding the joint spectral radius, for instance,
 the JSR toolbox in Matlab (Jungers 2023).
}
\references{
\itemize{
 \item C-T Chang and V.D. Blondel. 2013 . An experimental study of approximation algorithms for the joint spectral radius.
     \emph{Numerical algorithms}, \strong{64}, 181-202.
   \item Gripenberg, G. 1996. Computing the joint spectral radius. \emph{Linear Algebra and its Applications},
     234, 43–60.
   \item I.L. Kheifets, P.J. Saikkonen. 2020. Stationarity and ergodicity of Vector STAR models.
     \emph{Econometric Reviews}, \strong{39}:4, 407-414.
   \item R. Jungers (2023). The JSR toolbox (https://www.mathworks.com/matlabcentral/fileexchange/33202-the-jsr-toolbox),
      MATLAB Central File Exchange.
 }
}
\keyword{internal}
