% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointSpectralRadius.R
\name{d_lift}
\alias{d_lift}
\title{Calculate the dth induced matrix of a square matrix}
\usage{
d_lift(A, d)
}
\arguments{
\item{A}{a real square matrix}

\item{d}{a strictly positive integer specifying the order of the lift.}

\item{multisets}{all \eqn{d}-element multisets in \eqn{\lbrace 1,...,nrow(A) \rbrace} in
a lexicographic order (obtained with get_multisets_Cpp).}
}
\value{
Returns the dth induced matrix of A: a square matrix with \code{choose(nrow(A) + d - 1, d)}
  rows and columns.
}
\description{
\code{d_lift} the dth induced matrix of a square matrix
 by so-called d-lifting as described in Parrilo and Jadbabaie (2008).
}
\details{
Calculated the d-lifting using the formula presented in Parrilo and Jadbabaie (2008),
  Equation~(8).
}
\references{
\itemize{
   \item P.A. Parrilo, A. Jadbabaie. 2008. Approximation of the joint spectral
      radius using sum of squares. \emph{Linear Algebra and its Applications},
      \strong{428}, 2385-2402.
 }
}
\keyword{internal}
