% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlot.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Residual diagnostic plot for a STVAR model}
\usage{
diagnostic_plot(
  stvar,
  type = c("all", "series", "ac", "ch", "dist"),
  maxlag = 12
)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{type}{which type of diagnostic plot should be plotted?
\itemize{
  \item{\code{"all"} all below sequentially.}
  \item{\code{"series"} the residual time series.}
  \item{\code{"ac"} the residual autocorrelation and cross-correlation functions.}
  \item{\code{"ch"} the squared residual autocorrelation and cross-correlation functions.}
  \item{\code{"dist"} the residual histogram with theoretical density (dashed line) and QQ-plots.}
}}

\item{maxlag}{the maximum lag considered in types \code{"ac"} and \code{"ch"}.}
}
\description{
\code{diagnostic_plot} plots a multivariate residual diagnostic plot
  for either autocorrelation, conditional heteroskedasticity, or distribution,
  or simply draws the residual time series.
}
\details{
Auto- and cross-correlations (types \code{"ac"} and \code{"ch"}) are calculated with the function
 \code{acf} from the package \code{stats} and the plot method for class \code{'acf'} objects is employed.
}
\examples{
# FILL IN EXAMPLS
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
         \emph{Springer}.
   \item McElroy T. 2017. Computation of vector ARMA autocovariances.
         \emph{Statistics and Probability Letters}, \strong{124}, 92-96.
   \item References to the STVAR models TO BE INCLUDED.
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
 \code{\link{LR_test}}, \code{\link{Wald_test}}, \code{\link{cond_moment_plot}}
}
