% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticPlot.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Residual diagnostic plot for a STVAR model}
\usage{
diagnostic_plot(
  stvar,
  type = c("all", "series", "ac", "ch", "dist"),
  resid_type = c("standardized", "raw"),
  maxlag = 12
)
}
\arguments{
\item{stvar}{object of class \code{"stvar"}}

\item{type}{which type of diagnostic plot should be plotted?
\itemize{
  \item{\code{"all"} all below sequentially.}
  \item{\code{"series"} the residual time series.}
  \item{\code{"ac"} the residual autocorrelation and cross-correlation functions.}
  \item{\code{"ch"} the squared residual autocorrelation and cross-correlation functions.}
  \item{\code{"dist"} the residual histogram with theoretical density (dashed line) and QQ-plots.}
}}

\item{resid_type}{should standardized or raw residuals be used?}

\item{maxlag}{the maximum lag considered in types \code{"ac"} and \code{"ch"}.}
}
\description{
\code{diagnostic_plot} plots a multivariate residual diagnostic plot
  for either autocorrelation, conditional heteroskedasticity, or distribution,
  or simply draws the residual time series.
}
\details{
Auto- and cross-correlations (types \code{"ac"} and \code{"ch"}) are calculated with the function
 \code{acf} from the package \code{stats} and the plot method for class \code{'acf'} objects is employed.
 If \code{cond_dist == "Student"}, the estimate of the degrees of freedom parameter is used in theoretical
 densities and quantiles.
}
\examples{
# p=1, M=1, d=2, linear VAR
theta_112 <- c(0.649526, 0.066507, 0.288526, 0.021767, -0.144024, 0.897103,
  0.601786, -0.002945, 0.067224)
mod112 <- STVAR(data=gdpdef, p=1, M=1, params=theta_112)
diagnostic_plot(mod112, resid_type="standardized")
diagnostic_plot(mod112, resid_type="raw")

# p=1, M=2, d=2, relative dens weight function
theta_122relg <- c(0.734054, 0.225598, 0.705744, 0.187897, 0.259626, -0.000863,
  -0.3124, 0.505251, 0.298483, 0.030096, -0.176925, 0.838898, 0.310863, 0.007512,
  0.018244, 0.949533, -0.016941, 0.121403, 0.573269)
mod122 <- STVAR(data=gdpdef, p=1, M=2, params=theta_122relg)
diagnostic_plot(mod122, type="ac", resid_type="raw")
diagnostic_plot(mod122, type="ch", resid_type="standardized")
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
         \emph{Springer}.
   \item McElroy T. 2017. Computation of vector ARMA autocovariances.
         \emph{Statistics and Probability Letters}, \strong{124}, 92-96.
   \item References to the STVAR models TO BE INCLUDED.
 }
}
\seealso{
\code{\link{profile_logliks}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
 \code{\link{LR_test}}, \code{\link{Wald_test}}, \code{\link{Rao_test}}
}
