% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreEst.R
\name{fitSSTVAR}
\alias{fitSSTVAR}
\title{Maximum likelihood estimation of a structural STVAR model based on preliminary estimates from
  a reduced form model.}
\usage{
fitSSTVAR(
  stvar,
  new_identification = c("recursive", "heteroskedasticity"),
  new_B_constraints = NULL,
  maxit = 1000,
  maxit_robust = 1000,
  robust_method = c("none", "Nelder-Mead", "SANN"),
  calc_std_errors = TRUE
)
}
\arguments{
\item{stvar}{a an object of class \code{'stvar'}, created by, e.g., \code{fitSTVAR},
specifying a reduced form or a structural model}

\item{maxit}{the maximum number of iterations in the variable metric algorithm.}

\item{maxit_robust}{the maximum number of iterations on the first phase robust estimation, if employed.}

\item{robust_method}{Should some robust estimation method be used in the estimation before switching
to the gradient based variable metric algorithm? See details.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining the structural model
}
\description{
\code{fitSSTVAR} uses a variable metric algorithm to estimate a structural STVAR model
  based on preliminary estimates from a reduced form model.
}
\details{
When the structural model does not impose overidentifying constraints, it is directly
  obtained from the reduced form model, and estimation is not required. When overidentifying constraints
  are imposed, the model is estimated via ..

  Structural models can be provided in the argument \code{stvar} if overidentifying constraints should be
  imposed.

  Using a robust estimation method before switching to the variable metric can be useful if the initial
  estimates are not very close to the ML estimate of the structural model, as the variable metric algorithm
  (usually) converges to a nearby local maximum or saddle point. This is particularly the case when the imposed
  overidentifying restrictions are such that the unrestricted estimate is not close to satisfying them.

  Employs the estimation function \code{optim} from the package \code{stats} that implements the optimization
  algorithms. See \code{?optim} for the documentation on the
}
\examples{
\donttest{
## These are long running examples that use parallel computing!
## Running the below examples takes approximately FILL IN HOW MANY minutes
}
}
\references{
\itemize{
   \item TO BE FILLED IN
 }
}
\seealso{
\code{\link{fitSTVAR}}, \code{\link{STVAR}}, \code{\link[stats]{optim}}
}
