% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAINest.R
\name{fitSTVAR}
\alias{fitSTVAR}
\title{Two-phase maximum likelihood estimation of a reduced form smooth transition VAR model}
\usage{
fitSTVAR(
  data,
  p,
  M,
  weight_function = c("relative_dens", "logit"),
  cond_dist = c("Gaussian", "Student"),
  parametrization = c("intercept", "mean"),
  AR_constraints = NULL,
  mean_constraints = NULL,
  nrounds = (M + 1)^5,
  ncores = 2,
  maxit = 1000,
  seeds = NULL,
  print_res = TRUE,
  use_parallel = FALSE,
  filter_estimates = TRUE,
  ...
)
}
\arguments{
\item{data}{data a matrix or class \code{'ts'} object with \code{d>1} columns. Each column is taken to represent
a univariate time series. Missing values are not supported.}

\item{p}{a positive integer specifying the autoregressive order}

\item{M}{a positive integer specifying the number of regimes}

\item{weight_function}{what type of transition weights should be used?}

\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"} or \eqn{"Student"}.}

\item{parametrization}{\code{"intercept"} or \code{"mean"} determining whether the model is parametrized with intercept
parameters \eqn{\phi_{m,0}} or regime means \eqn{\mu_{m}}, m=1,...,M.}

\item{nrounds}{the number of estimation rounds that should be performed.}

\item{ncores}{the number CPU cores to be used in parallel computing.}

\item{maxit}{the maximum number of iterations in the variable metric algorithm.}

\item{seeds}{a length \code{nrounds} vector containing the random number generator seed
for each call to the genetic algorithm, or \code{NULL} for not initializing the seed.}

\item{print_res}{should summaries of estimation results be printed?}

\item{use_parallel}{employ parallel computing?}

\item{filter_estimates}{should the likely inappropriate estimates be filtered? See details.}

\item{...}{additional settings passed to the function \code{GAfit} employing the genetic algorithm.}
}
\value{
Returns an object of class \code{'stvar'} defining the estimated reduced form smooth transition VAR model.
}
\description{
\code{fitSTVAR} estimates a reduced form smooth transition VAR model in two phases:
  in the first phase, it uses a genetic algorithm to find starting values for a gradient based
  variable metric algorithm, which it then uses to finalize the estimation in the second phase.
  Parallel computing is utilized to perform multiple rounds of estimations in parallel.
}
\details{
If you wish to estimate a structural model, estimate first the reduced form model and then use the
 use the function FILL IN to estimate the structural model based on the estimated reduced form model.

 Because of complexity and high multimodality of the log-likelihood function, it is \strong{not certain}
 that the estimation algorithm will end up in the global maximum point. It is expected that most of the
 estimation rounds will end up in some local maximum or saddle point instead. Therefore, a (sometimes large)
 number of estimation rounds is required for reliable results. Because of the nature of the model,
 the estimation may fail especially in the cases where the number of regimes is chosen too large.

 The estimation process is computationally heavy and it might take considerably long time for large models with
 large number of observations. If the iteration limit \code{maxit} in the variable metric algorithm is reached,
 one can continue the estimation by iterating more with the function \code{iterate_more}. Alternatively, one may
 use the found estimates as starting values for the genetic algorithm and and employ another round of estimation
 (see \code{?GAfit} how to set up an initial population with the dot parameters).

 \strong{If the estimation algorithm fails to create an initial population for the genetic algorithm},
 it usually helps to scale the individual series so that the AR coefficients (of a VAR model) will be
 relative small, preferably less than one. Even if one is able to create an initial population, it should
 be preferred to scale the series so that most of the AR coefficients will not be very large, as the
 estimation algorithm works better with relatively small AR coefficients. If needed, another package can be used
 to fit linear VARs to the series to see which scaling of the series results in relatively small AR coefficients.

 \strong{Filtering inappropriate estimates:} If \code{filter_estimates == TRUE}, the code will automatically filter
 through estimates that it deems "inappropriate". That is, estimates that are not likely solutions of interest.
 Specifically, solutions that incorporate a near-singular error term covariance matrix (any eigenvalue less than \eqn{0.002})
 or transition weights such that they are close to zero for almost all \eqn{t} for at least one regime.
 You are welcome to set \code{filter_estimates=FALSE} and find the solutions of interest yourself by using the
 function \code{alt_stvar}.
}
\section{S3 methods}{

  The following S3 methods are supported for class \code{'stvar'}: \code{logLik}, \code{residuals}, \code{print}, \code{summary},
   \code{predict}, \code{simulate}, and \code{plot}. NONE OF THESE IS IMPLEMENTED YET!
}

\examples{
\donttest{
## These are long running examples that use parallel computing!
# Running all the below examples will take approximately FILL IN HOW MANY minutes.

# FILL IN
}
}
\references{
\itemize{
   \item FILL IN REFEENCES
 }
}
\seealso{
\code{\link{GAfit}}
}
