% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{get_alpha_mt}
\alias{get_alpha_mt}
\title{Get the transition weights alpha_mt}
\usage{
get_alpha_mt(
  data,
  Y2,
  p,
  M,
  d,
  weight_function,
  all_A,
  all_boldA,
  all_Omegas,
  weightpars,
  weightfun_pars,
  all_mu,
  epsilon,
  log_mvdvalues = NULL
)
}
\arguments{
\item{data}{data a matrix or class \code{'ts'} object with \code{d>1} columns. Each column is taken to represent
a univariate time series. Missing values are not supported.}

\item{Y2}{the data arranged as obtained from \code{reform_data(data, p)} but excluding the last row}

\item{p}{a positive integer specifying the autoregressive order}

\item{M}{a positive integer specifying the number of regimes}

\item{d}{the number of time series in the system, i.e., the dimension}

\item{weight_function}{what type of transition weights should be used? See the vignette for details about the weight functions.}

\item{all_A}{4D array containing all coefficient matrices \eqn{A_{m,i}}, obtained from \code{pick_allA}.}

\item{all_boldA}{3D array containing the \eqn{((dp)x(dp))} "bold A" (companion form) matrices of each regime,
obtained from \code{form_boldA}. Will be computed if not given.}

\item{weightpars}{numerical vector containing the transition weight parameters, obtained from \code{pick_weightpars}.}

\item{weightfun_pars}{\describe{
  \item{If \code{weight_function == "relative_dens"}:}{Not used.}
  \item{If \code{weight_function == "logit"}:}{a list of two elements: \describe{
    \item{The first element \code{$vars}:}{a numeric vector containing the variables that should used as switching variables
      in the weight function in an increasing order, i.e., a vector with unique elements in \eqn{\lbrace 1,...,d \rbrace}.}
    \item{The second element \code{$lags}:}{an integer in \eqn{\lbrace 1,...,p \rbrace} specifying the number of lags to be
     used in the weight function.}
  }}
}}

\item{all_mu}{an \eqn{(d \times M)} matrix containing the unconditional regime-specific means}

\item{epsilon}{the smallest number such that its exponent is wont classified as numerically zero
(around \code{-698} is used).}

\item{log_mvdvalues}{a \eqn{T x M} matrix containing log multivariate normal densities (can be used with
relative dens weight function only)}
}
\value{
Returns the mixing weights a \eqn{(T x M)} matrix, so that the t:th row is for the time point t
  and m:th column is for the regime m.
}
\description{
\code{get_alpha_mt} computes the transition weights.
}
\details{
Note that we index the time series as \eqn{-p+1,...,0,1,...,T}.
}
\references{
\itemize{
   \item TO BE FILLED IN
 }
 @keywords internal
}
\keyword{internal}
