% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_multisets_Cpp}
\alias{get_multisets_Cpp}
\title{Generate all k-element multisets}
\usage{
get_multisets_Cpp(n, k, N)
}
\arguments{
\item{n}{positive integer specifying the value n in \eqn{\lbrace 1,...,n \rbrace}.}

\item{k}{positive integer specifying the size of the multisets.}

\item{N}{the binomial coefficient \code{choose(n + k - 1, k)} (calculated in R wrapper)}
}
\value{
a numeric matrix with N rows and d columns, containing each multiset in each row, ordered
  lexicographically from top to bottom.
}
\description{
Generates all k-element multisets in \eqn{\lbrace 1,...,n \rbrace}
  ordered in a lexicographic order.
}
\details{
This function is used in the least squares estimation
}
\keyword{internal}
