% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histDecomp.R, R/plotMethods.R, R/printMethods.R
\name{hist_decomp}
\alias{hist_decomp}
\alias{plot.histdecomp}
\alias{print.histdecomp}
\title{Compute historical decompositions for structural STVAR models.}
\usage{
hist_decomp(stvar)

\method{plot}{histdecomp}(x, ..., plot_by_shock = FALSE, which_to_plot)

\method{print}{histdecomp}(x, ..., digits = 3, which_vars, which_indices)
}
\arguments{
\item{x}{object of class \code{'histdecomp'} generated by the function \code{hist_decomp}.}

\item{...}{currently not in use.}

\item{plot_by_shock}{should the historical decompositions by plotted so that there
is one figure for each shock (rather than one figure for each variable)?}

\item{which_to_plot}{a numeric vector with the indices of the variables or shocks
(depending on the argument \code{plot_by_shock}) to be plotted. The default is to plot
all of them.}

\item{digits}{how many significant digits to print?}

\item{which_vars}{a numeric vector specifying the variables to print. The default is that
all the variables are printed.}

\item{which_indices}{a numeric vector specifying the time period indices to print. The default
is that all the time periods are printed.}
}
\value{
Returns a class \code{'histdecomp'} list with the following elements:
  \describe{
    \item{init_cond_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the initial conditions to the movements of
     the variables at each time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{steady_state_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the steady state component to the movements of
     the variables at each time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{shock_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the shocks to the movements of the variables at each
     time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{contributions_of_shocks}{A 3D array of size \eqn{(T \times d \times d)} containing the cumulative contributions of the shocks to the
     movements of the variables at each time point; the element \code{t, i1, i2} giving the contribution of the shock \code{i1} to the variable
     \code{i2} at the time \code{t}.}
    \item{stvar}{The original STVAR model object.}
 }

Returns the input object \code{x} invisibly.
}
\description{
\code{hist_decomp} compute historical decompositions for structural STVAR models.
}
\details{
The historical decomposition quantifies the cumulative effects the shocks to the movements of
  the variables (see, e.g., Kilian and Lütkepohl, 2017, Section~4.3) The historical decompositions are
  computed as described in Wong (2018). Note that due to the effect of the "initial conditions" and the
  "steady state component", which are not attributed to the shocks, the cumulative effects of the shocks
  do not sum to the observed time series.
}
\section{Functions}{
\itemize{
\item \code{plot(histdecomp)}: plot method

\item \code{print(histdecomp)}: print method

}}
\examples{
 ## FILL IN
}
\seealso{
\code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{linear_IRF}}, \code{\link{fitSSTVAR}}
 \itemize{
   \item Kilian L., Lütkepohl H. 2017. Structural Vector Autoregressive Analysis. 1st edition.
     \emph{Cambridge University Press}, Cambridge.
   \item Wong H. 2018. Historical decomposition for nonlinear vector autoregressive models.
     \emph{CAMA Working Paper No. 62/2017, available as SSRN:3057759}
 }
}
