% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histDecomp.R
\name{hist_decomp}
\alias{hist_decomp}
\title{Compute historical decompositions for structural STVAR models.}
\usage{
hist_decomp(stvar)
}
\value{
Returns a class \code{'histdecomp'} list with the following elements:
  \describe{
    \item{init_cond_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the initial conditions to the movements of
     the variables at each time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{steady_state_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the steady state component to the movements of
     the variables at each time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{shock_comp}{A matrix of size \eqn{(T \times d)} containing the contributions of the shocks to the movements of the variables at each
     time point; the element \code{t, i} giving the contribution at the time \code{t} on the variable \code{i}.}
    \item{contributions_of_shocks}{A 3D array of size \eqn{(T \times d \times d)} containing the cumulative contributions of the shocks to the
     movements of the variables at each time point; the element \code{t, i1, i2} giving the contribution of the shock \code{i1} to the variable
     \code{i2} at the time \code{t}.}
    \item{stvar}{The original STVAR model object.}
 }
}
\description{
\code{hist_decomp} compute historical decompositions for structural STVAR models.
}
\details{
The historical decomposition quantifies the cumulative effects the shocks to the movements of
  the variables (see, e.g., Kilian and Lütkepohl, 2017, Section~4.3) The historical decompositions are
  computed as described in Wong (2018). Note that due to the effect of the "initial conditions" and the
  "steady state component", which are not attributed to the shocks, the cumulative effects of the shocks
  do not sum to the observed time series.
}
\examples{
 ## FILL IN
}
\seealso{
\code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{linear_IRF}}, \code{\link{fitSSTVAR}}
 \itemize{
   \item Kilian L., Lütkepohl H. 2017. Structural Vector Autoregressive Analysis. 1st edition.
     \emph{Cambridge University Press}, Cambridge.
   \item Wong H. 2018. Historical decomposition for nonlinear vector autoregressive models.
     \emph{CAMA Working Paper No. 62/2017, available as SSRN:3057759}
 }
}
