% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{in_paramspace}
\alias{in_paramspace}
\title{Determine whether the parameter vector is in the parameter space}
\usage{
in_paramspace(
  p,
  M,
  d,
  weight_function,
  cond_dist,
  all_boldA,
  all_Omegas,
  weightpars,
  df,
  stab_tol = 0.001,
  posdef_tol = 1e-08,
  df_tol = 1e-08
)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order}

\item{M}{a positive integer specifying the number of regimes}

\item{d}{the number of time series in the system, i.e., the dimension}

\item{weight_function}{what type of transition weights should be used? See the vignette for details about the weight functions.}

\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"} or \eqn{"Student"}.}

\item{all_boldA}{3D array containing the \eqn{((dp)x(dp))} "bold A" (companion form) matrices of each regime,
obtained from \code{form_boldA}. Will be computed if not given.}

\item{weightpars}{numerical vector containing the transition weight parameters, obtained from \code{pick_weightpars}.}

\item{stab_tol}{numerical tolerance for the stability condition of each regime: if the "bold A" matrix of any regime
has eigenvalues larger that \code{1 - stab_tol} the parameter vector will be labeled as not in the parameter space.
Note that if the tolerance is too small, numerical evaluation of the log-likelihood might fail.}

\item{posdef_tol}{numerical tolerance for positive definiteness of the regime-specific covariance matrices: if
the error term covariance matrix of any regime has eigenvalues smaller than this, the model is classified
as not satisfying positive definiteness assumption. Note that if the tolerance is too small, numerical
evaluation of the log-likelihood might fail and cause error.}

\item{df_tol}{the parameter vector is considered to be outside the parameter space the degrees of
freedom parameters is not larger than \code{2 + df_tol}.}

\item{all_Omega}{3D array containing all covariance matrices \eqn{\Omega_{m}}, obtained from \code{pick_Omegas}.}
}
\value{
Returns \code{TRUE} if the given parameter values are in the parameter space and \code{FALSE} otherwise.
  This function does NOT consider the identifiability condition!
}
\description{
\code{in_paramspace} checks whether the parameter vector is in the parameter
  space.
}
\details{
The parameter vector in the argument \code{params} should be unconstrained and reduced form.
}
\references{
\itemize{
   \item TO BE FILLED IN
 }
 @keywords internal
}
