% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreEst.R
\name{iterate_more}
\alias{iterate_more}
\title{Maximum likelihood estimation of a reduced form or structural STVAR model based on preliminary estimates}
\usage{
iterate_more(stvar, maxit = 100, calc_std_errors = TRUE)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'}, created by, e.g., \code{fitSTVAR}.}

\item{maxit}{the maximum number of iterations in the variable metric algorithm.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining the estimated model
}
\description{
\code{iterate_more} uses a variable metric algorithm to estimate a reduced form or structural STVAR model
 (object of class \code{'stvar'}) based on preliminary estimates.
}
\details{
The purpose of \code{iterate_more} is to provide a simple and convenient tool to finalize
  the estimation when the maximum number of iterations is reached when estimating a STVAR model
  with the main estimation function \code{fitSTVAR} or \code{fitSSTVAR}.
}
\examples{
\donttest{
## These are long running examples that use parallel computing!
## Running the below examples takes approximately FILL IN HOW MANY minutes

# p=1, M=2, d=2 relative_dens Gaussian STVAR, only 5 iterations of
# the variable matrix algorithm.
fit12 <- fitSTVAR(gdpdef, p=1, M=2, nrounds=1, seeds=1, maxit=5, use_parallel=FALSE)
fit12

# Iterate more:
fit12 <- iterate_more(fit12)
fit12
}
}
\references{
\itemize{
   \item TO BE FILLED IN
 }
}
\seealso{
\code{\link{fitSTVAR}}, \code{\link{STVAR}}, \code{\link[stats]{optim}}
}
