% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreEst.R
\name{iterate_more}
\alias{iterate_more}
\title{Maximum likelihood estimation of a reduced form or structural STVAR model based on preliminary estimates}
\usage{
iterate_more(stvar, maxit = 100, calc_std_errors = TRUE)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'}, created by, e.g., \code{fitSTVAR} or \code{fitSSTVAR}.}

\item{maxit}{the maximum number of iterations in the variable metric algorithm.}

\item{calc_std_errors}{should approximate standard errors be calculated?}
}
\value{
Returns an object of class \code{'stvar'} defining the estimated model
}
\description{
\code{iterate_more} uses a variable metric algorithm to estimate a reduced form or structural STVAR model
 (object of class \code{'stvar'}) based on preliminary estimates.
}
\details{
The purpose of \code{iterate_more} is to provide a simple and convenient tool to finalize
  the estimation when the maximum number of iterations is reached when estimating a STVAR model
  with the main estimation function \code{fitSTVAR} or \code{fitSSTVAR}.
}
\examples{
\donttest{
## These are long running examples that take approximately 20 seconds to run.

# Estimate two-regime Gaussian STVAR p=1 model with the weighted relative stationary densities
# of the regimes as the transition weight function, but only 5 iterations of the variable matrix
# algorithm:
fit12 <- fitSTVAR(gdpdef, p=1, M=2, nrounds=1, seeds=1, ncores=1, maxit=5)

# The iteration limit was reached, so the estimate is not local maximum.
# The gradient of the log-likelihood function:
get_foc(fit12) # Not close to zero!

# So, we run more iterations of the variable metric algorithm:
fit12 <- iterate_more(fit12)

# The gradient of the log-likelihood function after iterating more:
get_foc(fit12) # Close to zero!
}
}
\references{
\itemize{
   \item Anderson H., Vahid F. 1998. Testing multiple equation systems for common nonlinear components.
     \emph{Journal of Econometrics}, \strong{84}:1, 1-36.
   \item Hubrich K., Teräsvirta. T. 2013. Thresholds and Smooth Transitions in Vector Autoregressive Models.
     \emph{CREATES Research Paper 2013-18, Aarhus University.}
   \item Lanne M., Virolainen S. 2024. A Gaussian smooth transition vector autoregressive model:
      An application to the macroeconomic effects of severe weather shocks. Unpublished working
      paper, available as \url{https://doi.org/10.48550/arXiv.2403.14216}.
   \item Kheifets I.L., Saikkonen P.J. 2020. Stationarity and ergodicity of Vector STAR models.
     \emph{Econometric Reviews}, \strong{39}:4, 407-414.
   \item Tsay R. 1998. Testing and Modeling Multivariate Threshold Models.
     \emph{Journal of the American Statistical Association}, \strong{93}:443, 1188-1202.
   \item Virolainen S. 2024. Identification by non-Gaussianity in structural threshold and
      smooth transition vector autoregressive models. Unpublished working
      paper, available in ArXiv.
 }
}
\seealso{
\code{\link{fitSTVAR}}, \code{\link{STVAR}}, \code{\link[stats]{optim}},
 \code{\link{swap_B_signs}}, \code{\link{reorder_B_columns}}
}
