% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearIRF.R
\name{linear_IRF}
\alias{linear_IRF}
\title{Estimate linear impulse response function based on a single regime of a structural STVAR model.}
\usage{
linear_IRF(
  stvar,
  N = 30,
  regime = 1,
  which_cumulative = numeric(0),
  scale = NULL,
  ci = NULL,
  bootstrap_reps = 100,
  ncores = 2,
  nrounds = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'} defining a structural or reduced form
STVAR model. For a reduced form model, the shocks are automatically identified by
the lower triangular Cholesky decomposition.}

\item{N}{a positive integer specifying the horizon how far ahead should the
linear impulse responses be calculated.}

\item{regime}{Based on which regime the linear IRF should be calculated?
An integer in \eqn{1,...,M}.}

\item{which_cumulative}{a numeric vector with values in \eqn{1,...,d}
(\code{d=ncol(data)}) specifying which the variables for which the linear impulse
responses should be cumulative. Default is none.}

\item{scale}{should the linear IRFs to some of the shocks be scaled so that they
correspond to a specific instantaneous response of some specific
variable? Provide a length three vector where the shock of interest
is given in the first element (an integer in \eqn{1,...,d}), the variable of
interest is given in the second element (an integer in \eqn{1,...,d}), and
its instantaneous response in the third element (a non-zero real number).
If the linear IRFs of multiple shocks should be scaled, provide a matrix which has one
column for each of the shocks with the columns being the length three vectors described above.}

\item{ci}{a real number in \eqn{(0, 1)} specifying the confidence level of the
confidence intervals calculated via a fixed-design wild residual bootstrap method.
Available only for models that impose linear autoregressive dynamics
(excluding changes in the volatility regime).}

\item{bootstrap_reps}{the number of bootstrap repetitions for estimating confidence bounds.}

\item{ncores}{the number of CPU cores to be used in parallel computing when bootstrapping confidence bounds.}

\item{nrounds}{on how many estimation rounds should each bootstrap estimation be based on?
Does not have to be very large since initial estimates used are based on the initially fitted model.
Larger number of rounds gives more reliable results but is computationally more demanding.}

\item{...}{parameters passed to the plot method \code{plot.irf} that plots
the results.}

\item{seeds}{a numeric vector of length \code{bootstrap_reps} initializing the seed for the random
generator for each bootstrap replication.}
}
\value{
Returns a class \code{'irf'} list with the linear IRFs in ... FILL IN!
}
\description{
\code{linear_IRF} estimates linear impulse response function based on a single regime
  of a structural STVAR model.
}
\details{
If the autoregressive dynamics of the model are linear (i.e., either M == 1 or mean and AR parameters
  are constrained identical across the regimes), confidence bounds can be calculated based on a fixed-design
  wild residual bootstrap method. We employ the method described in Herwartz and L端tkepohl (2014); see also
  the relevant chapters in Kilian and L端tkepohl (2017).
}
\examples{
 # FILL IN
}
\references{
\itemize{
   \item Herwartz H. and L端tkepohl H. 2014. Structural vector autoregressions with Markov switching:
     Combining conventional with statistical identification of shocks. \emph{Journal of Econometrics},
     183, pp. 104-116.
   \item Kilian L. and L端tkepohl H. 2017. Structural Vectors Autoregressive Analysis.
         \emph{Cambridge University Press}, Cambridge.
 }
}
\seealso{
\code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
  \code{\link{reorder_W_columns}}, \code{\link{swap_W_signs}}
}
