% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{n_params}
\alias{n_params}
\title{Calculate the number of (freely estimaed) parameters in the model}
\usage{
n_params(
  p,
  M,
  d,
  weight_function = c("relative_dens", "logit"),
  cond_dist = c("Gaussian", "Student"),
  identification = c("reduced_form", "recursive", "heteroskedasticity"),
  AR_constraints = NULL,
  mean_constraints = NULL,
  B_constraints = NULL
)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order}

\item{M}{a positive integer specifying the number of regimes}

\item{d}{the number of time series in the system, i.e., the dimension}

\item{weight_function}{what type of transition weights should be used?}

\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"} or \eqn{"Student"}.}

\item{identification}{is it reduced form model or an identified structural model; if the latter, how is it identified?}
}
\value{
Returns the number of parameters in the parameter vector of the specified model.
}
\description{
\code{n_params} calculates the number of (freely estimaed) parameters in the model.
}
\section{Warning}{

 No argument checks!
}

\references{
\itemize{
   \item TO BE FILLED IN
 }
 @keywords internal
}
\keyword{internal}
