% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChecks.R
\name{n_params}
\alias{n_params}
\title{Calculate the number of (freely estimaed) parameters in the model}
\usage{
n_params(
  p,
  M,
  d,
  weight_function = c("relative_dens", "logit"),
  cond_dist = c("Gaussian", "Student"),
  identification = c("reduced_form", "recursive", "heteroskedasticity"),
  AR_constraints = NULL,
  mean_constraints = NULL,
  B_constraints = NULL
)
}
\arguments{
\item{p}{a positive integer specifying the autoregressive order}

\item{M}{a positive integer specifying the number of regimes}

\item{d}{the number of time series in the system, i.e., the dimension}

\item{weight_function}{what type of transition weights should be used?}

\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"} or \eqn{"Student"}.}

\item{identification}{is it reduced form model or an identified structural model; if the latter, how is it identified?}

\item{AR_constraints}{a size \eqn{(Mpd^2 x q)} constraint matrix \eqn{C} specifying general linear constraints
to the autoregressive parameters. The constraints are of the form
\eqn{(\varphi_{1},...,\varphi_{M}) = C\psi}, where \eqn{\varphi_{m} = (vec(A_{m,1}),...,vec(A_{m,p})) \ (pd^2 x 1),\ m=1,...,M},
contains the coefficient matrices and \eqn{\psi} \eqn{(q x 1)} contains the related parameters.
For example, to restrict the AR-parameters to be the identical across the regimes, set \eqn{C =}
[\code{I:...:I}]' \eqn{(Mpd^2 x pd^2)} where \code{I = diag(p*d^2)}.}

\item{mean_constraints}{Restrict the mean parameters of some regimes to be identical? Provide a list of numeric vectors
such that each numeric vector contains the regimes that should share the common mean parameters. For instance, if
\code{M=3}, the argument \code{list(1, 2:3)} restricts the mean parameters of the second and third regime to be
identical but the first regime has freely estimated (unconditional) mean. Ignore or set to \code{NULL} if mean parameters
should not be restricted to be the same among any regimes. \strong{This constraint is available only for mean parametrized models;
that is, when \code{parametrization="mean"}.}}

\item{B_constraints}{NOT YET IMPLEMENTED!}
}
\value{
Returns the number of parameters in the parameter vector of the specified model.
}
\description{
\code{n_params} calculates the number of (freely estimaed) parameters in the model.
}
\section{Warning}{

 No argument checks!
}

\references{
\itemize{
   \item TO BE FILLED IN
 }
 @keywords internal
}
\keyword{internal}
