% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagTests.R
\name{Portmanteau_test}
\alias{Portmanteau_test}
\title{Perform adjusted Portmanteau test for a STVAR model}
\usage{
Portmanteau_test(stvar, nlags = 20, which_test = c("autocorr", "het.sked"))
}
\arguments{
\item{stvar}{an object of class \code{'stvar'} generated by \code{fitSTVAR} or \code{STVAR}.}

\item{nlags}{a strictly positive integer specifying the number of lags to be tested.}

\item{type}{should test for remaining autocorrelation or heteroskedasticity be calculated?}
}
\value{
A list with class "hypotest" containing the test results and arguments used to calculate the test.
}
\description{
\code{Portmanteau_test} performs adjusted Portmanteau test for remaining autocorrelation
  (or heteroskedasticity) in the residuals of a STVAR model.
}
\details{
The implemented adjusted Portmanteau test is based on Lütkepohl (2005), Section 4.4.3.
  When testing for remaining heteroskedasticity, the Portmanteau test is applied to squared
  standardized residuals. Note that the validity of the heteroskedasticity test requires that the
  residuals are not autocorrelated.
}
\examples{
# FILL IN
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
           \emph{Springer}.
  }
}
\seealso{
\code{\link{LR_test}}, \code{\link{Rao_test}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
  \code{\link{diagnostic_plot}}, \code{\link{profile_logliks}},
}
