% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stvar}
\alias{predict.stvar}
\title{Predict method for class 'stvar' objects}
\usage{
\method{predict}{stvar}(
  object,
  ...,
  nsteps,
  nsim = 1000,
  pi = c(0.95, 0.8),
  pred_type = c("mean", "median")
)
}
\arguments{
\item{object}{an object of class \code{'stvar'}.}

\item{...}{currently not in use.}

\item{nsteps}{how many steps ahead should be predicted?}

\item{nsim}{to how many independent simulations should the forecast be based on?}

\item{pi}{a numeric vector specifying the confidence levels of the prediction intervals.}

\item{pred_type}{should the pointforecast be based on sample "median" or "mean"?}
}
\value{
Returns a class '\code{stvarpred}' object containing, among the specifications,...
 \describe{
   \item{$pred}{Point forecasts}
   \item{$pred_ints}{Prediction intervals, as \code{[, , d]}.}
   \item{$trans_pred}{Point forecasts for the transition weights}
   \item{$trans_pred_ints}{Individual prediction intervals for transition weights, as \code{[, , m]}, m=1,..,M.}
 }
}
\description{
\code{predict.stvar} is a predict method for class \code{'stvar'} objects. The forecasts
  are computed by simulating multiple sample paths of the future observations and using the
  sample medians or means as point forecasts and empirical quantiles as prediction intervals.
}
\examples{
# FILL IN EXAMPLES
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
         \emph{Springer}.
   \item McElroy T. 2017. Computation of vector ARMA autocovariances.
         \emph{Statistics and Probability Letters}, \strong{124}, 92-96.
   \item References to the STVAR models TO BE INCLUDED.
 }
}
\seealso{
\code{\link{simulate.stvar}}
}
