% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileLogliks.R
\name{profile_logliks}
\alias{profile_logliks}
\title{Plot profile log-likelihood functions about the estimates}
\usage{
profile_logliks(
  stvar,
  which_pars,
  scale = 0.02,
  nrows,
  ncols,
  precision = 200,
  stab_tol = 0.001,
  posdef_tol = 1e-08,
  distpar_tol = 1e-08,
  weightpar_tol = 1e-08
)
}
\arguments{
\item{stvar}{an object of class \code{'stvar'}, created by, e.g., \code{fitSTVAR}.}

\item{which_pars}{the profile log-likelihood function of which parameters should be plotted? An integer
vector specifying the positions of the parameters in the parameter vector. The parameter vector has the
form...}

\item{scale}{a numeric scalar specifying the interval plotted for each estimate:
the estimate plus-minus \code{abs(scale*estimate)}.}

\item{nrows}{how many rows should be in the plot-matrix? The default is \code{max(ceiling(log2(length(which_pars)) - 1), 1)}.}

\item{ncols}{how many columns should be in the plot-matrix? The default is \code{ceiling(length(which_pars)/nrows)}.
Note that \code{nrows*ncols} should not be smaller than the length of \code{which_pars}.}

\item{precision}{at how many points should each profile log-likelihood function be evaluated at?}

\item{stab_tol}{numerical tolerance for stability of condition of the regimes: if the "bold A" matrix of any regime
has eigenvalues larger that \code{1 - stat_tol} the parameter is considered to be outside the parameter space.
Note that if tolerance is too small, numerical evaluation of the log-likelihood might fail and cause error.}

\item{posdef_tol}{numerical tolerance for positive definiteness of the error term covariance matrices: if
the error term covariance matrix of any regime has eigenvalues smaller than this, the parameter is considered
to be outside the parameter space. Note that if the tolerance is too small, numerical evaluation of the
log-likelihood might fail and cause error.}

\item{distpar_tol}{the parameter vector is considered to be outside the parameter space if the degrees of
freedom parameters is not larger than \code{2 + distpar_tol} (applies only if \code{cond_dist="Student"}).}

\item{weightpar_tol}{numerical tolerance for weight parameters being in the parameter space. Values closer to
to the border of the parameter space than this are considered to be "outside" the parameter space.}
}
\value{
Only plots to a graphical device and doesn't return anything.
}
\description{
\code{profile_logliks} plots profile log-likelihood functions about the estimates.
}
\details{
When the number of parameters is large, it might be better to plot a smaller number of profile
 log-likelihood functions at a time using the argument \code{which_pars}.

The red vertical line points the estimate.
}
\examples{
\donttest{
# Running all the below examples takes approximately FILL IN HOW MANY MINUTES

# FILL IN
}
}
\references{
\itemize{
   \item Lütkepohl H. 2005. New Introduction to Multiple Time Series Analysis,
         \emph{Springer}.
   \item McElroy T. 2017. Computation of vector ARMA autocovariances.
         \emph{Statistics and Probability Letters}, \strong{124}, 92-96.
   \item References to the STVAR models TO BE INCLUDED.
 }
}
\seealso{
\code{\link{get_foc}}, \code{\link{get_soc}}, \code{\link{diagnostic_plot}}
}
