% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{random_distpars}
\alias{random_distpars}
\title{Create random distribution parameter values}
\usage{
random_distpars(d, cond_dist)
}
\arguments{
\item{cond_dist}{specifies the conditional distribution of the model as \code{"Gaussian"}, \code{"Student"}, or \code{"ind_Student"},
where the latest is the Student's \eqn{t} distribution with independent components.}
}
\value{
Returns a numeric vector ...
  \describe{
    \item{If \code{cond_dist == "Gaussian"}:}{of length zero.}
    \item{If \code{cond_dist == "Student"}:}{of length one containing a df param strictly larger than two.}
    \item{If \code{cond_dist == "ind_Student"}:}{of length d containing a df params strictly larger than two.}
  }
}
\description{
\code{random_distpars} generates random distribution parameter values
}
\keyword{internal}
