% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{random_impactmat}
\alias{random_impactmat}
\title{Create random VAR model impact matrix}
\usage{
random_impactmat(d, B_scale, m)
}
\arguments{
\item{B_scale}{a size \eqn{(d \times 1)} strictly positive vector specifying the mean and variability of the
random impact matrices in random mutations. In each regime, the mean of the error term covariance matrix
implied by the random impact matrix will be \code{diag(B_scale)}, and the variability increases with \code{B_scale}.
Default in \code{GAfit} is \code{var(stats::ar(data[,i], order.max=10)$resid, na.rm=TRUE), i=1,...,d}.
This argument is ignored if \code{cond_dist != "ind_Student"}.}

\item{m}{which regime? Regime 1 impact matrix is constrained so the elements in its first row
are in a decreasing ordering and the diagonal elements are strictly positive.}
}
\value{
Returns a \eqn{(d^2 \times 1)} vector containing the vectorized impact matrix \eqn{B}.
}
\description{
\code{random_impactmat} generates random VAR model \eqn{(dxd)} impact matrix \eqn{B}
  from a custom distribution related to the Wishart distribution.
}
\keyword{internal}
