% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{random_impactmat}
\alias{random_impactmat}
\title{Create random VAR model impact matrix}
\usage{
random_impactmat(d, B_scale)
}
\arguments{
\item{B_scale}{a size \eqn{(d \times 1)} strictly positive vector specifying the mean and variability of the
random impact matrices in random mutations. The mean of the random impact matrix will be
\code{diag(omega_scale)}, and the variability increases with \code{omega_scale}. Default in \code{GAfit}
is \code{sd(stats::ar(data[,i], order.max=10)$resid, na.rm=TRUE), i=1,...,d}. This argument is ignored if
\code{cond_dist != "ind_Student"}.}
}
\value{
Returns a \eqn{(d^2 \times 1)} vector containing vectorized impact matrix \eqn{B}.
}
\description{
\code{random_covmat} generates random VAR model \eqn{(dxd)} impact matrix \eqn{B}
  with its diagonal entries normalized to be positive. The impact matrix is generaged from a
  custom distribution related to the Wishart distribution.
}
\keyword{internal}
