% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{smart_df}
\alias{smart_df}
\title{Create random degrees of freedom parameter values close to given values}
\usage{
smart_df(df, accuracy)
}
\arguments{
\item{df}{the old degrees of freedom parameters (of all regimes)}

\item{accuracy}{a positive real number adjusting how close to the given degrees of freedom parameters
the returned df should be.}
}
\value{
a numeric vector containing the degrees of freedom parameter values
}
\description{
\code{smart_df} generates random degrees of freedom parameter values
  close to given values.
}
\keyword{internal}
