% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{smart_distpars}
\alias{smart_distpars}
\title{Create random distribution parameter values close to given values}
\usage{
smart_distpars(distpars, accuracy, cond_dist)
}
\arguments{
\item{distpars}{the old distribution parameters (of all regimes)}

\item{accuracy}{a positive real number adjusting how close to the given distribution parameters
the returned values should be.}
}
\value{
Returns a numeric vector ...
  \describe{
    \item{If \code{cond_dist == "Gaussian"}:}{of length zero.}
    \item{If \code{cond_dist == "Student"}:}{of length one containing a degrees of freedom parameter value
     (strictly larger than two).}
    \item{If \code{cond_dist == "ind_Student"}:}{of length d containing a degrees of freedom parameter values
     (strictly larger than two).}
    \item{If \code{cond_dist == "ind_skewed_t"}:}{of length 2d containing df params strictly larger than two in the first d
          elements and skewness params strictly between -1 and 1 in the rest d elements.}
  }
}
\description{
\code{smart_distpars} generates random degrees of freedom parameter values
  close to given values.
}
\keyword{internal}
