% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstvars.R
\docType{package}
\name{sstvars-package}
\alias{sstvars}
\alias{sstvars-package}
\title{sstvars: toolkit for structural smooth transition vector autoregressive models}
\description{
\code{sstvars} is a package for reduced form and structural smooth transition vector
  autoregressive models with focus on the structural analysis. It provides functions for ...

  Some of the code is taken from the R package gmvarkit (also created by the author of this package)
  but is copied rather than using the code in gmvarkit directly in order to allow independent development.

  The readme file is a good place to start and the vignette might be useful too.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saviviro/sstvars}
  \item Report bugs at \url{https://github.com/saviviro/sstvars/issues}
}

}
\author{
you <savi.virolainen@helsinki.fi>
}
