% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usacpu}
\alias{usacpu}
\title{A monthly  U.S. data covering the period from 1987:4 to 2024:2 (443 observations) and consisting six variables.
First, the climate policy uncertainty index (CPUI) (Gavridiilis, 2021), which is a news based measure of climate policy uncertainty.
Second, the economic policy uncertainty index (EPUI), which is a news based measure of economic policy uncertainty.
Third, the log-difference of real indsitrial production index (IPI).
Fourth, the log-difference of the consumer price index (CPI).
Fifth, the log-difference of the producer price index (PPI).
Sixth, an interest rate variable, which is the effective federal funds rate that is replaced by the
the Wu and Xia (2016) shadow rate during zero-lower-bound periods. The Wu and Xia (2016) shadow rate is not bounded by the zero
lower bound and also quantifies unconventional monetary policy measures, while it closely follows the federal funds rate when the
zero lower bound does not bind. This is the dataset used in Virolainen (2024)}
\format{
A numeric matrix of class \code{'ts'} with 443 rows and 4 columns with one time series in each column:
\describe{
  \item{First column (CPUI):}{The climate policy uncertainty index, \url{https://www.policyuncertainty.com/climate_uncertainty.html}.}
  \item{Second column (EPUI):}{The economic policy uncertainty index, \url{https://www.policyuncertainty.com/us_monthly.html}.}
  \item{Third column (IPI):}{The log-difference of real indsitrial production index, \url{https://fred.stlouisfed.org/series/INDPRO}.}
  \item{Fourth column (CPI):}{The log-difference of the consumer price index, \url{https://fred.stlouisfed.org/series/CPIAUCSL}.}
  \item{Fifth column (PPI):}{The log-difference of the producer price index, \url{https://fred.stlouisfed.org/series/PPIACO}.}
  \item{Sixth column (RATE):}{The Federal funds rate from 1954Q3 to 2008Q2 and after that the Wu and Xia (2016) shadow rate,
   \url{https://fred.stlouisfed.org/series/FEDFUNDS}, \url{https://www.atlantafed.org/cqer/research/wu-xia-shadow-federal-funds-rate}.}
}
}
\source{
The Federal Reserve Bank of St. Louis database and the Federal Reserve Bank of Atlanta's website
}
\usage{
usacpu
}
\description{
A monthly  U.S. data covering the period from 1987:4 to 2024:2 (443 observations) and consisting six variables.
First, the climate policy uncertainty index (CPUI) (Gavridiilis, 2021), which is a news based measure of climate policy uncertainty.
Second, the economic policy uncertainty index (EPUI), which is a news based measure of economic policy uncertainty.
Third, the log-difference of real indsitrial production index (IPI).
Fourth, the log-difference of the consumer price index (CPI).
Fifth, the log-difference of the producer price index (PPI).
Sixth, an interest rate variable, which is the effective federal funds rate that is replaced by the
the Wu and Xia (2016) shadow rate during zero-lower-bound periods. The Wu and Xia (2016) shadow rate is not bounded by the zero
lower bound and also quantifies unconventional monetary policy measures, while it closely follows the federal funds rate when the
zero lower bound does not bind. This is the dataset used in Virolainen (2024)
}
\references{
\itemize{
   \item K. Gavriilidis, 2021. Measuring climate policy uncertainty. \url{https://www.ssrn.com/abstract=3847388}.
   \item Federal Reserve Bank of Chicago. 2023. Monthly GDP Growth Rate Data. \url{https://www.chicagofed.org/publications/bbki/index}.
   \item Virolainen S. 2024. Identification by non-Gaussianity in structural threshold and smooth transition vector
     autoregressive models. Unpublished working paper, available as arXiv.2404.19707.
   \item Wu J. and Xia F. 2016. Measuring the macroeconomic impact of monetary policy at the zero lower bound.
     \emph{Journal of Money, Credit and Banking}, 48(2-3): 253-291.
 }
}
\keyword{datasets}
